Combine one if else statement

+++ src/main/java/com/bbn/hadder/listener/InviteLinkListener.java
@@ -44,21 +44,19 @@ public class InviteLinkListener extends ListenerAdapter {
                         e.printStackTrace();
                     }
                 }
-            } else if (event.getMessage().getContentRaw().contains("discordapp.com/invite")) {
-                if (!event.getMember().hasPermission(Permission.ADMINISTRATOR) && rethink.getInviteDetection(event.getGuild().getId())) {
-                    String split = event.getMessage().getContentRaw().split("discordapp.com/invite/", 10)[1];
-                    String invite = split.split(" ")[0];
-                    OkHttpClient client = new OkHttpClient();
-                    Request request = new Request.Builder().url("https://canary.discordapp.com/api/v6/invites/" + invite).build();
-                    try {
-                        Response response = client.newCall(request).execute();
-                        JSONObject json = new JSONObject(response.body().string());
-                        if (!json.toString().contains("\"message\":")) {
-                            event.getMessage().delete().reason("Invite Link detected").queue();
-                        }
-                    } catch (Exception e) {
-                        e.printStackTrace();
+            } else if (event.getMessage().getContentRaw().contains("discordapp.com/invite") && !event.getMember().hasPermission(Permission.ADMINISTRATOR) && rethink.getInviteDetection(event.getGuild().getId())) {
+                String split = event.getMessage().getContentRaw().split("discordapp.com/invite/", 10)[1];
+                String invite = split.split(" ")[0];
+                OkHttpClient client = new OkHttpClient();
+                Request request = new Request.Builder().url("https://canary.discordapp.com/api/v6/invites/" + invite).build();
+                try {
+                    Response response = client.newCall(request).execute();
+                    JSONObject json = new JSONObject(response.body().string());
+                    if (!json.toString().contains("\"message\":")) {
+                        event.getMessage().delete().reason("Invite Link detected").queue();
                     }
+                } catch (Exception e) {
+                    e.printStackTrace();
                 }
             }
         }
